backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
requests==2.32.3
numpy==1.23.5
pandas==2.0.3
joblib==1.4.2
Pillow==9.5.0
tensorflow-cpu==2.12.0
keras==2.12.0
